+++ src/grepros/__init__.py
@@ -6,10 +6,10 @@ Released under the BSD License.
 
 @author      Erki Suurjaak
 @created     31.10.2021
-@modified    06.01.2022
+@modified    07.01.2022
 ------------------------------------------------------------------------------
 """
 ## @namespace grepros
 __title__        = "grepros"
-__version__      = "0.4.1.dev15"
-__version_info__ = (0, 4, 1, "dev15")
+__version__      = "0.4.1.dev16"
+__version_info__ = (0, 4, 1, "dev16")