+++ CHANGELOG.rst
@@ -6,10 +6,11 @@ Changelog for package grepros
 -------------------
 * refactor database and SQL sinks onto a common base
 * allow specifying dialect options in Postgres/SQLite output
-* allow overriding table and view names in SQL-related output
+* allow overriding table and view names in SQL dialects
 * drop meta-table from Postgres output
 * drop default value from --publish-prefix option
 * fix error in parsing subtypes from message definitions
+* fix converting ROS2 temporal messages to seconds/nanoseconds
 
 0.4.0 (2021-12-26)
 -------------------